requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
boto3==1.34.91
python-dotenv==1.0.1
asyncpg==0.29.0
nest_asyncio==1.6.0
supabase-py==2.0.3
aiosmtplib==2.0.2
python-multipart==0.0.9
pydantic==2.7.1
